Pipelining Commit #4:
Implemented a resolution for an ID/EX conflict (when the target of a reg
write of a preceding instruction is the same as one of the current
instruction's source registers. (Example: instruction 1 writes 0x1 to
register $t0, instruction 2 writes $t0 + $t0 into $t. W/o resolution,
after 2 finishes, $t0 will store 0. With resolution, it will store 0x2))

TODO:
- implement conflict resolution for ID/MEM conflicts (when a reg is
written to in the MEM phase, requires a delay that allows for the reg
to be updated before continuing the current instruction.)